Change price to int

+++ TruffleMarketApi/Services/Item/ItemBatchModel.cs
@@ -2,7 +2,7 @@
 {
     public class ItemBatchModel
     {
-        public double MaxTotalPrice { get; set; }
+        public int MaxTotalPrice { get; set; }
         public int MinTotalWeight { get; set; }
         public string Truffle {get; set;}
     }
